etl/requirements.txt
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
